successfully send http request with restTemplate and successfully parse response json
