+++ pom.xml
@@ -170,6 +170,10 @@
   		    <artifactId>log4j</artifactId>
   		    <version>1.2.17</version>
   	    </dependency>
+		<dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+		</dependency>
  
 	</dependencies>
