+++ src/main/java/com/revature/Driver.java
@@ -3,6 +3,9 @@ package com.revature;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+import com.revature.beans.geolocationResponse.GeoLocationResponse;
 
 import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -10,7 +13,6 @@ import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
-import com.revature.services.DistanceService;
 
 
 /**
@@ -33,6 +35,11 @@ public class Driver {
 	public static void main(String[] args) {
 		SpringApplication.run(Driver.class, args); 
 		
+		//test making request to GeoLocationAPI and parse response
+		RestTemplate restTemplate = new RestTemplate();
+		String url = "https://maps.googleapis.com/maps/api/geocode/json?address=1001+s+center+st,+Arlington+TX+76010&key=AIzaSyDdU7KMs5IfYmqJXmYZc8MzL_8fN3z5Hec";
+		GeoLocationResponse gr = restTemplate.getForObject(url, GeoLocationResponse.class);
+		System.out.println(gr.getResults().get(0).getFormatted_address());
 	}
 	
 	/**
@@ -64,5 +71,6 @@ public class Driver {
 			.apiInfo(apiInfo());
 	}
 	
-
+	
+	
 }
