+++ src/main/java/com/revature/beans/geolocationResponse/GeoLocationResponse.java
@@ -1,0 +1,74 @@
+package com.revature.beans.geolocationResponse;
+
+import java.io.Serializable;
+import java.util.ArrayList;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class GeoLocationResponse implements Serializable {
+
+	private static final long serialVersionUID = 1L;
+	
+	private ArrayList<Result> results;
+	private String status;
+	
+	public GeoLocationResponse() {
+		super();
+	}
+
+	public ArrayList<Result> getResults() {
+		return results;
+	}
+
+	public void setResults(ArrayList<Result> results) {
+		this.results = results;
+	}
+
+	public String getStatus() {
+		return status;
+	}
+
+	public void setStatus(String status) {
+		this.status = status;
+	}
+
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + ((results == null) ? 0 : results.hashCode());
+		result = prime * result + ((status == null) ? 0 : status.hashCode());
+		return result;
+	}
+
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj)
+			return true;
+		if (obj == null)
+			return false;
+		if (getClass() != obj.getClass())
+			return false;
+		GeoLocationResponse other = (GeoLocationResponse) obj;
+		if (results == null) {
+			if (other.results != null)
+				return false;
+		} else if (!results.equals(other.results))
+			return false;
+		if (status == null) {
+			if (other.status != null)
+				return false;
+		} else if (!status.equals(other.status))
+			return false;
+		return true;
+	}
+
+	@Override
+	public String toString() {
+		return "GeoLocationResponse [results=" + results + ", status=" + status + "]";
+	}
+	
+}
+
+
