+++ src/main/java/com/revature/beans/geolocationResponse/Location.java
@@ -1,0 +1,28 @@
+package com.revature.beans.geolocationResponse;
+
+import java.io.Serializable;
+
+public class Location implements Serializable {
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
+	private String lat;
+	private String lng;
+	public Location() {
+		super();
+	}
+	public String getLat() {
+		return lat;
+	}
+	public void setLat(String lat) {
+		this.lat = lat;
+	}
+	public String getLng() {
+		return lng;
+	}
+	public void setLng(String lng) {
+		this.lng = lng;
+	}
+	
+}
